refactor:save()方法移至基类,增加账单具体model
